requirements.txt
Django==2.0.7
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
psycopg2==2.7.5
pytz==2018.5
